Ran cargo clipy on both files, changed accordingly
By running clippy on the files, there were some useless things such as
push_str(&format) stuff which could be simplified to use the write!() or
the writeln!() macro which work like write!(var, "something{}",
somethinghere), which pushes the formated string (using the last two
parameters), into the var (first parameter). This is useful as it avoids
an extra heap allocation.