Added multithreading (std::thread and rayon)
I created two types of multithreading possibilities: using std::tread
and just spawning one for every file (adding them to a vector and then
at the end joining them all together - if this is not done they might
get killed when the function finishes running but the threads have not
yet completed what they needed to do). The trouble here is controlling
the amount of threads, as that needs to be implemented somehow.

The other method was done via rayon, and likely the one I will keep
using for now. We just added a ".par_iter()" method onto the list of
files and then worked on them as we would before, and it seems to work
nicely. We also added a few lines of code to control the amount of
threads in the program (default 3, otherwise take from input).